+++ pydidas/widgets/misc/__init__.py
@@ -29,6 +29,7 @@ __all__ = []
 
 # import __all__ items from modules:
 from .line_edit_with_icon import *
+from .select_points_in_image import *
 from .show_information_for_result import *
 from .read_only_text_widget import *
 
@@ -39,12 +40,17 @@ from . import line_edit_with_icon
 __all__.extend(line_edit_with_icon.__all__)
 del line_edit_with_icon
 
-from . import read_only_text_widget
+from . import select_points_in_image
 
-__all__.extend(read_only_text_widget.__all__)
-del read_only_text_widget
+__all__.extend(select_points_in_image.__all__)
+del select_points_in_image
 
 from . import show_information_for_result
 
 __all__.extend(show_information_for_result.__all__)
 del show_information_for_result
+
+from . import read_only_text_widget
+
+__all__.extend(read_only_text_widget.__all__)
+del read_only_text_widget